Added lines which produce a TMY vector.
Created a matrix which ndicates the year each month came from.

Added a vector which stores the FS of the months selected
for the composite year.

Adjusted the indeces so that the script will run a less than 12 month
year. This would be useful it we want to isolate a particular month or
season.